Improved address parsing and error output
This commit addresses an issue in Bitcoin Core where an incorrect
bech32 chain prefix would lead to the address being treated as entirely
invalid for both base58 and bech32 decoding, without providing the user
with an appropriate error message.

When a user inputs the incorrect address DecodeDestination now informs
the user of the input and expected chain prefixes as well as the network
name string.

There were other improvements to error displays
    - bech32.cpp : Added too short of data portion instead of just
      "invalid separator position"
    - key_io.cpp : Now accounts for more cases of valid char sets but
      imporper checksums. Also some gramatical mistakes were fixed.

This does pass the extended functional test suite.

Refactor for GetChainTypeString()

Co-authored-by: D++ <82842780+dplusplus1024@users.noreply.github.com>